- Add Python bindings build system with nanobind and scikit-build-core
- Add core Python bindings: Api, Session, EventQueue, enums, events, exceptions
- Fix EventIterator queue closure detection and deduplicate session_state_str
- Add Python bindings for DeviceInfo, FullInfo, and VariableDefinitions
- Fix TelemetryDefinitions::end() returning begin() instead of end()
- Add Python bindings for Telemetry, FFB Pipeline, and Clock
- Add bit variable read support to Python bindings
- Fix Phase 4 Python binding issues found during review
- Add Python bindings for LED Control, SimData read/write
- Fix Phase 5 Python binding issues found during review
- Add Python example scripts mirroring C++ examples
- Add CI python-wheel job and cibuildwheel config
- Add generation scripts, templates, and public data files
- Wire generation into CMake and remove static files
- Guard libeddsa subdirectory against duplicate target
- Guard sc-api-core-protocol_types against duplicate target
- Add missing device.h include for DeviceSessionId
- Fix compiler warnings and clang-tidy issues in sc-api-core
- Remove unused TelemtryGroup::enabled_ private variable
- Add .claude to .gitignore

+++ core/src/session.cpp
@@ -420,7 +420,7 @@ ResultCode Session::registerToControl(uint32_t control_flags, const std::string&
     p_->receive_buffer.resize(k_rx_buffer_size);
 
     std::size_t written_bytes = asio::write(p_->main_socket, asio::buffer(ptr, size));
-    if (size != written_bytes) {
+    if (static_cast<std::size_t>(size) != written_bytes) {
         p_->main_socket.close();
         return ResultCode::error_cannot_connect;
     }
@@ -442,7 +442,7 @@ ResultCode Session::registerToControl(uint32_t control_flags, const std::string&
                 return ResultCode::error_protocol;
             }
             needed_bytes = total_document_size;
-            if (read_bytes == total_document_size) {
+            if (read_bytes == static_cast<std::size_t>(total_document_size)) {
                 break;
             }
         }
